region_cross event: turn off physics and reposition inside original region
...before posting the region_cross event, so the script can do things like
llSetRot() without hanging the server, as the server would try to do
sog.AbsolutePosition = sog.AbsolutePosition (using the outside-of-region
coordinates) before setting the rotation and cause an infinite loop trying
to position the object outside the region.  So change eliminates that
possibility by putting the object back in the original region with physics
off so sog.AbsolutePosition returns an inside-region coordinate.