GridUserService patch for finding user's home location
http://opensimulator.org/mantis/view.php?id=6986 :

I received a report from a user in a grid that every time they logged in and
asked to be sent to their home location (in land they owned), or to their last
location they would always wind up at a different location in the grid. The
other location was always the same each time. If the user specified a region
name and location at the login prompt they would arrive where they wanted.